refactor: 関数移動
